1st task separeted from main.py
